feat: moved the evidence use case into separate files
